refactor(nextjs): Move to client components and metadata added
